move _select_order_coefs to utils
