+++ pyfixest/utils.py
@@ -1,3 +1,5 @@
+import re
+
 import numpy as np
 import pandas as pd
 from formulaic import model_matrix
@@ -255,3 +257,44 @@ def simultaneous_crit_val(C: np.ndarray, S: int, alpha: float = 0.05) -> float:
     p = C.shape[0]
     tmaxs = np.max(np.abs(msqrt(C) @ np.random.randn(p, S).reshape(p, S)), axis=0)
     return np.quantile(tmaxs, 1 - alpha)
+
+
+def _select_order_coefs(res: pd.DataFrame, keep: list, drop: list):
+    """
+    Select and order the coefficients based on the pattern.
+
+    Parameters
+    ----------
+    res: pd.DataFrame
+        The DataFrame to be ordered.
+    keep: list
+        Refer to the `keep` parameter in the `etable` function.
+    drop: list
+        Refer to the `drop` parameter in the `etable` function.
+
+    Returns
+    -------
+    res: pd.DataFrame
+        The ordered DataFrame.
+    """
+    coefs = list(res.index)
+    coef_order = [] if keep else coefs[:]  # Store matched coefs
+    for pattern in keep:
+        _coefs = []  # Store remaining coefs
+        for coef in coefs:
+            if re.findall(pattern, coef):
+                coef_order.append(coef)
+            else:
+                _coefs.append(coef)
+        coefs = _coefs
+
+    for pattern in drop:
+        _coefs = []
+        for (
+            coef
+        ) in coef_order:  # Remove previously matched coefs that match the drop pattern
+            if not re.findall(pattern, coef):
+                _coefs.append(coef)
+        coef_order = _coefs
+
+    return res.loc[coef_order]
